Overhaul of SaveFile loading; ObjectListView
Large overhaul of the loading of the SaveFile. Should be a little more organized and intuitive now.
Put in some ground work for the new ObjectListView control. Everything on the second Advanced tab will replace what is on the first. This is being done because ObjectListView is much more powerful than the standard ListView component.
Some of the code is very messy right now due to the transition to ObjectListView. Many things are broken (necessarily) as of this commit.